Update psf_selection and catalog_seed_image to not depend on psf_library; separate segment_psfs module
