+++ mirage/psf/psf_selection.py
@@ -20,7 +20,7 @@ Use
 
     This module can be imported and called as such:
     ::
-        from mirage.psf import psf_selction
+        from mirage.psf import psf_selection
         library = psf_selection.get_gridded_psf_library('nircam', 'nrcb1',
                                                         'f200w', 'clear',
                                                         'predicted', 0,
@@ -33,8 +33,6 @@ from glob import glob
 import os
 
 from astropy.io import fits
-import numpy as np
-import pysiaf
 from webbpsf.utils import to_griddedpsfmodel
 
 from mirage.utils.constants import NIRISS_PUPIL_WHEEL_FILTERS
@@ -186,56 +184,6 @@ def get_gridded_psf_library(instrument, detector, filtername, pupilname, wavefro
     return library
 
 
-def get_gridded_segment_psf_library_list(instrument, detector, filtername,
-                                         library_path, pupilname="CLEAR"):
-    """Find the filenames for the appropriate gridded segment PSF libraries and
-    read them into griddedPSFModel objects
-
-    Parameters
-    ----------
-    instrument : str
-        Name of instrument the PSFs are from
-
-    detector : str
-        Name of the detector within ```instrument```
-
-    filtername : str
-        Name of filter used for PSF library creation
-
-    library_path : str
-        Path pointing to the location of the PSF library
-
-    pupilname : str, optional
-        Name of pupil wheel element used for PSF library creation. Default is "CLEAR".
-
-    Returns:
-    --------
-    libraries : list of photutils.griddedPSFModel
-        List of object containing segment PSF libraries
-
-    """
-    library_list = get_segment_library_list(instrument, detector, filtername, library_path, pupil=pupilname)
-
-    print("Segment PSFs will be generated using:")
-    for filename in library_list:
-        print(os.path.basename(filename))
-
-    libraries = []
-    for filename in library_list:
-        with fits.open(filename) as hdulist:
-            hdr = hdulist[0].header
-            d = hdulist[0].data
-
-        data = d[0][0]
-        phdu = fits.PrimaryHDU(data, header=hdr)
-        hdulist = fits.HDUList(phdu)
-
-        lib_model = to_griddedpsfmodel(hdulist)
-        libraries.append(lib_model)
-
-    return libraries
-
-
 def get_library_file(instrument, detector, filt, pupil, wfe, wfe_group,
                      library_path, wings=False, segment_id=None):
     """Given an instrument and filter name along with the path of
@@ -279,7 +227,6 @@ def get_library_file(instrument, detector, filt, pupil, wfe, wfe_group,
     psf_files = glob(os.path.join(library_path, '*.fits'))
 
     # Create a dictionary of header information for all PSF library files
-    # psf_table = {}
     matches = []
 
     instrument = instrument.upper()
@@ -294,17 +241,7 @@ def get_library_file(instrument, detector, filt, pupil, wfe, wfe_group,
 
         # Compare the header entries to the user input
         file_inst = header['INSTRUME'].upper()
-
-        if segment_id is None:
-            try:
-                file_det = header['DETECTOR'].upper()
-            except KeyError:
-                file_det = header['DET_NAME'].upper()
-            det_match = file_det == detector
-        else:
-            det_keys = [k for k in header.keys() if 'DETNAME' in k]
-            dets = [header[k] for k in det_keys]
-            det_match = detector in dets
+        file_det = header['DETECTOR'].upper()
         file_filt = header['FILTER'].upper()
 
         try:
@@ -345,7 +282,7 @@ def get_library_file(instrument, detector, filt, pupil, wfe, wfe_group,
 
         # Evaluate if the file matches the given parameters
         match = (file_inst == instrument
-                 and det_match
+                 and file_det == detector
                  and file_filt == filt
                  and file_pupil == pupil)
         if not wings and segment_id is None:
@@ -366,50 +303,6 @@ def get_library_file(instrument, detector, filt, pupil, wfe, wfe_group,
         raise ValueError("No PSF library file found matching requested parameters.")
     elif len(matches) > 1:
         raise ValueError("More than one PSF library file matches requested parameters: {}".format(matches))
-
-
-def get_segment_library_list(instrument, detector, filt,
-                             library_path, pupil='CLEAR'):
-    """Given an instrument and filter name along with the path of
-    the PSF library, find the appropriate 18 segment PSF library files.
-
-    Parameters
-    -----------
-    instrument : str
-        Name of instrument the PSFs are from
-
-    detector : str
-        Name of the detector within ```instrument```
-
-    filt : str
-        Name of filter used for PSF library creation
-
-    library_path : str
-        Path pointing to the location of the PSF library
-
-    pupil : str, optional
-        Name of pupil wheel element used for PSF library creation. Default is
-        'CLEAR'.
-
-    segment_id : int or None, optional
-        If specified, returns a segment PSF library file and denotes the ID
-        of the mirror segment
-
-    Returns
-    --------
-    library_list : list
-        List of the names of the segment PSF library files for the instrument
-        and filter name
-    """
-    library_list = []
-    for seg_id in np.arange(1, 19):
-         segment_file = get_library_file(
-             instrument, detector, filt, pupil, '', 0, library_path,
-             segment_id=seg_id
-         )
-         library_list.append(segment_file)
-
-    return library_list
 
 
 def get_psf_wings(instrument, detector, filtername, pupilname, wavefront_error, wavefront_error_group,
@@ -493,74 +386,3 @@ def get_psf_wings(instrument, detector, filtername, pupilname, wavefront_error, 
                    "These must be even."))
             raise ValueError
     return psf_wing
-
-
-def get_segment_offset(segment_number, detector, library_list):
-    """Convert vectors coordinates in the local segment control
-    coordinate system to NIRCam detector X and Y coordinates,
-    at least proportionally, in order to calculate the location
-    of the segment PSFs on the given detector.
-
-    Parameters
-    ----------
-    segment : int
-        Segment ID, i.e 3
-    detector : str
-        Name of NIRCam detector
-    library_list : list
-        List of the names of the segment PSF library files
-
-    Returns
-    -------
-    x_displacement
-        The shift of the segment PSF in NIRCam SW x pixels
-    y_displacement
-        The shift of the segment PSF in NIRCam SW y pixels
-    """
-
-    # Verify that the segment number in the header matches the index
-    seg_index = int(segment_number) - 1
-    header = fits.getheader(library_list[seg_index])
-
-    assert int(header['SEGID']) == int(segment_number), \
-        "Uh-oh. The segment ID of the library does not match the requested " \
-        "segment. The library_list was not assembled correctly."
-    xtilt = header['XTILT']
-    ytilt = header['YTILT']
-    segment = header['SEGNAME'][:2]
-
-    # These conversion factors were empirically calculated by measuring the
-    # relation between tilt and the pixel displacement
-    tilt_to_pixel_slope = 13.4
-    tilt_to_pixel_intercept = 0
-
-    control_xaxis_rotations = {
-        'A1': 180, 'A2': 120, 'A3': 60, 'A4': 0, 'A5': -60,
-        'A6': -120, 'B1': 0, 'C1': 60, 'B2': -60, 'C2': 0,
-        'B3': -120, 'C3': -60, 'B4': -180, 'C4': -120,
-        'B5': -240, 'C5': -180, 'B6': -300, 'C6': -240
-    }
-
-    x_rot = control_xaxis_rotations[segment]  # degrees
-    x_rot_rad = x_rot * np.pi / 180  # radians
-
-    # Note that y is defined as the x component and x is defined as the y component.
-    # This is because "xtilt" moves the PSF in the y direction, and vice versa.
-    tilt_onto_y = (xtilt * np.cos(x_rot_rad)) - (ytilt * np.sin(x_rot_rad))
-    tilt_onto_x = (xtilt * np.sin(x_rot_rad)) + (ytilt * np.cos(x_rot_rad))
-
-    # TODO: IS THE SLOPE DIFFERENT FOR LW DETECTORS????
-    x_displacement = -(tilt_onto_x * tilt_to_pixel_slope) + tilt_to_pixel_intercept  # pixels
-    y_displacement = -(tilt_onto_y * tilt_to_pixel_slope) + tilt_to_pixel_intercept  # pixels
-
-    # Get the appropriate pixel scale from pysiaf
-    siaf = pysiaf.Siaf('nircam')
-    aperture = siaf['NRC{}_FULL'.format(detector[-2:].upper())]
-    nircam_x_pixel_scale = aperture.XSciScale  # arcsec/pixel
-    nircam_y_pixel_scale = aperture.YSciScale  # arcsec/pixel
-
-    # Convert the pixel displacement into angle
-    x_arcsec = x_displacement * nircam_x_pixel_scale  # arcsec
-    y_arcsec = y_displacement * nircam_y_pixel_scale  # arcsec
-
-    return x_arcsec, y_arcsec
